requirements.txt
streamlit==1.46.0
mysql-connector-python
fpdf
requests
pandas
mysql-connector-python
tabulate
fpdf
pandas
